requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.2.0
python-dotenv==1.0.1
drf-spectacular==0.27.0
django-filter==23.5
redis==5.0.1
gunicorn==21.2.0 